# Conflicts:
#	microservices/clients/contest-client/src/main/java/org/xcolab/client/contest/ContestClient.java
#	microservices/clients/contest-client/src/main/java/org/xcolab/client/contest/PlanTemplateClient.java
#	microservices/clients/contest-client/src/main/java/org/xcolab/client/contest/pojo/phases/AbstractContestPhase.java
#	microservices/clients/proposals-client/src/main/java/org/xcolab/client/proposals/PointsDistributionConfigurationClient.java
#	microservices/clients/proposals-client/src/main/java/org/xcolab/client/proposals/ProposalsClient.java
#	microservices/clients/proposals-client/src/main/java/org/xcolab/client/proposals/pojo/Proposal.java
#	portlets/contestmanagement-portlet/src/main/java/org/xcolab/portlets/contestmanagement/beans/ContestDescriptionBean.java
#	portlets/contestmanagement-portlet/src/main/java/org/xcolab/portlets/contestmanagement/wrappers/ProposalTemplateWrapper.java
#	portlets/contestmanagement-portlet/src/main/java/org/xcolab/portlets/contestmanagement/wrappers/SectionDefinitionWrapper.java
#	portlets/proposals-portlet/src/main/java/org/xcolab/portlets/proposals/utils/ProposalPickerFilter.java
#	portlets/proposals-portlet/src/main/java/org/xcolab/portlets/proposals/utils/ProposalPickerFilterUtil.java
#	services/plansProposalsFacade/plansProposalsFacade-portlet-service/src/main/java/org/xcolab/utils/OntologyTermToFocusAreaMapper.java

+++ microservices/services/proposals-service/src/main/java/org/xcolab/service/proposal/domain/proposalvote/ProposalVoteDaoImpl.java
@@ -100,9 +100,9 @@ public class ProposalVoteDaoImpl implements ProposalVoteDao {
     }
 
     @Override
-    public int delete(Long proposalId, Long contestPhaseId) {
+    public int delete(Long memberId, Long contestPhaseId) {
         return dslContext.deleteFrom(PROPOSAL_VOTE)
-                .where(PROPOSAL_VOTE.PROPOSAL_ID.eq(proposalId))
+                .where(PROPOSAL_VOTE.USER_ID.eq(memberId))
                 .and(PROPOSAL_VOTE.CONTEST_PHASE_ID.eq(contestPhaseId))
                 .execute();
     }
